Created minimal working interface with GiNaC
Note that the header for the exposed library must not contain
references to any nonstandard (according to MATLAB)
C++ features. That is, we should only expose
"vanilla" C++ functionality, e.g., no custom classes.

Hence, it's fine to use additional libraries
(as we did here with GiNaC), so long as
all of the GiNaC-related code is "under the hood".